+++ commons/util/src/main/java/net/automatalib/common/util/collection/IterableUtil.java
@@ -16,6 +16,9 @@
 package net.automatalib.common.util.collection;
 
 import java.util.Collection;
+import java.util.Collections;
+import java.util.Iterator;
+import java.util.List;
 import java.util.function.Function;
 import java.util.stream.Stream;
 
@@ -23,6 +26,68 @@ public final class IterableUtil {
 
     private IterableUtil() {
         // prevent instantiation
+    }
+
+    public static <T> Iterable<List<T>> allTuples(Iterable<? extends T> domain, int length) {
+        return allTuples(domain, length, length);
+    }
+
+    /**
+     * Returns an iterator that iterates over all tuples of the given source domain whose length (dimension) is within
+     * the specified range. Each intermediate combination of elements is computed lazily.
+     * <p>
+     * <b>Note:</b> Subsequent calls to the returned iterator's {@link Iterator#next() next()} method return a
+     * reference to the same list, and only update the contents of the list. If you plan to reuse intermediate results,
+     * you'll need to explicitly copy them.
+     *
+     * @param domain
+     *         the iterables for the source domains
+     * @param minLength
+     *         the minimal length of the tuple
+     * @param maxLength
+     *         the maximum length of the tuple
+     * @param <T>
+     *         type of elements
+     *
+     * @return an iterator that iterates over all tuples of the given source domain whose length (dimension) is within
+     * the specified range
+     */
+    public static <T> Iterable<List<T>> allTuples(Iterable<? extends T> domain, int minLength, int maxLength) {
+        // Check if domain is empty
+        // If it is, then the empty tuple (if not excluded by minLength > 0) is still part of the result
+        // Otherwise, the result is empty
+        if (!domain.iterator().hasNext()) {
+            if (minLength == 0) {
+                return Collections.singletonList(Collections.emptyList());
+            }
+            return Collections.emptyList();
+        }
+
+        return () -> new AllTuplesIterator<>(domain, minLength, maxLength);
+    }
+
+    /**
+     * Returns an iterator that iterates over the cartesian product of its given source domains. Each intermediate
+     * combination of elements is computed lazily.
+     * <p>
+     * <b>Note:</b> Subsequent calls to the returned iterator's {@link Iterator#next() next()} method return a
+     * reference to the same list, and only update the contents of the list. If you plan to reuse intermediate results,
+     * you'll need to explicitly copy them.
+     *
+     * @param iterables
+     *         the iterables for the source domains
+     * @param <T>
+     *         type of elements
+     *
+     * @return an iterator that iterates over the cartesian product of its given source domains
+     */
+    @SafeVarargs
+    public static <T> Iterable<List<T>> cartesianProduct(Iterable<T>... iterables) {
+        if (iterables.length == 0) {
+            return Collections.singletonList(Collections.emptyList());
+        }
+
+        return () -> new AllCombinationsIterator<>(iterables);
     }
 
     /**