+++ core/src/main/java/net/automatalib/automaton/base/AbstractCompact.java
@@ -28,7 +28,7 @@ import net.automatalib.automaton.MutableDeterministic.FullIntAbstraction;
 import net.automatalib.automaton.UniversalFiniteAlphabetAutomaton;
 import net.automatalib.automaton.concept.StateIDs;
 import net.automatalib.automaton.concept.StateLocalInput;
-import net.automatalib.common.util.collection.CollectionsUtil;
+import net.automatalib.common.util.collection.CollectionUtil;
 import org.checkerframework.checker.initialization.qual.UnknownInitialization;
 import org.checkerframework.checker.nullness.qual.Nullable;
 
@@ -79,7 +79,7 @@ public abstract class AbstractCompact<I, T, SP, TP> implements MutableAutomaton<
 
     @Override
     public Collection<Integer> getStates() {
-        return CollectionsUtil.intRange(0, numStates);
+        return CollectionUtil.intRange(0, numStates);
     }
 
     @Override